Moving checkGlobalL0Node to shared folder
