+++ .github/action_scripts/check_clusters/dag.js
@@ -1,14 +1,4 @@
-const { clusterCheck } = require("./shared")
-
-const checkGlobalL0Node = async () => {
-  const infos = [
-    {
-      name: 'Global L0',
-      baseUrl: 'http://localhost:9000'
-    }
-  ];
-  await clusterCheck(infos, true, 'Global L0', 1, true);
-};
+const { clusterCheck, checkGlobalL0Node } = require("./shared")
 
 const checkDAGL1Node = async () => {
   const infos = [
@@ -27,7 +17,6 @@ const checkDAGL1Node = async () => {
   ];
   await clusterCheck(infos, false, 'DAG L1', 3, true);
 };
-
 
 const main = async () => {
   await checkGlobalL0Node();
